Web-Application architecture/requirements.txt
openapi-python-client==0.11.2
fastapi==0.68.2
uvicorn==0.15.0
pydantic==1.9.0
openapi_client
apispec_fromfile
requests
flask
werkzeug.security
itsdangerous
sqlite3
